# docs/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "A modern Go-based authentication system with magic link functionality using Clean Architecture principles.\nThis API provides secure authentication mechanisms including magic link authentication, JWT token management, and user profile management.",
        "title": "Trading Alchemist API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Trading Alchemist Team",
            "email": "team@tradingalchemist.dev"
        },
        "license": {
            "name": "MIT",
            "url": "http://opensource.org/licenses/MIT"
        },
        "version": "1.0.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/conversations": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves a paginated list of active conversations for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Get user's conversations",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "Number of conversations to return",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "Offset for pagination",
                        "name": "offset",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversations retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/trading-alchemist_internal_application_chat.ConversationSummaryResponse"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Creates a new chat session for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Create a new conversation",
                "parameters": [
                    {
                        "description": "Conversation creation request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.CreateConversationRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Conversation created successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.ConversationDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid request body or validation error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{id}": {
            "get": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Retrieves the full details of a single conversation, including its messages, for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Get conversation details",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation details retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.ConversationDetailResponse"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Invalid conversation ID format",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not own this conversation",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conversation not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Archives a conversation for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Archive conversation",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation archived successfully",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid conversation ID format",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not own this conversation",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conversation not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{id}/messages": {
            "post": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Sends a message to a conversation and streams the LLM's response back using Server-Sent Events (SSE).",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "text/plain"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Post a message and get a streaming response",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Message content",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.PostMessageRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "text/event-stream response",
                        "schema": {
                            "type": "string"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or ID format",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not own this conversation",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conversation not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/conversations/{id}/title": {
            "put": {
                "security": [
                    {
                        "Bearer": []
                    }
                ],
                "description": "Updates the title of a conversation for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Chat"
                ],
                "summary": "Update conversation title",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Conversation ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Title update request",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_application_chat.UpdateConversationTitleRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Conversation title updated successfully",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Invalid request body or ID format",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "403": {
                        "description": "Forbidden - User does not own this conversation",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Conversation not found",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "trading-alchemist_internal_application_chat.ConversationSummaryResponse": {
            "type": "object",
            "properties": {
                "id": {
                    "type": "string"
                },
                "last_message_at": {
                    "type": "string"
                },
                "model_id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
...
        "trading-alchemist_internal_application_chat.CreateConversationRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "model_name": {
                    "description": "e.g., \"gpt-4o-mini\". Defaults to a system-wide default if not provided.",
                    "type": "string"
                },
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "trading-alchemist_internal_application_chat.UpdateConversationTitleRequest": {
            "type": "object",
            "required": [
                "title"
            ],
            "properties": {
                "title": {
                    "type": "string",
                    "maxLength": 255,
                    "minLength": 1
                }
            }
        },
...
        "trading-alchemist_internal_presentation_responses.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "description": "Error details",
                    "allOf": [
                        {
                            "$ref": "#/definitions/trading-alchemist_internal_presentation_responses.ErrorDetail"
                        }
                    ]
                },
                "success": {
                    "description": "Success indicator (always false for errors)\nexample: false",
                    "type": "boolean"
                }
            }
        },
        "trading-alchemist_internal_presentation_responses.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {
                    "description": "Response data"
                },
                "message": {
                    "description": "Optional success message\nexample: Operation completed successfully",
                    "type": "string"
                },
                "success": {
                    "description": "Success indicator\nexample: true",
                    "type": "boolean"
                }
            }
        }
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "JWT token. Usage: \"Bearer {token}\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}